Initial Storage class with reading and writing
The SRAM LDAP does not provide posixAccount attributes and when the
destination LDAP does, these attributes need to be generated. The
Storage class keeps track of these attributes and their values. If a
value does not exist, it can be generated and stored in the Storage
class.

The initial version of this class supported reading and writing of data
only. The generation of attribute values need to be implemented.